Flask API/requirements.txt
alembic==1.16.2
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
fuzzywuzzy==0.18.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.1.0
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
requests==2.32.4
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
